Adds permissions and Tenant service.
* Adds spatie/laravel-permissions
* Adds namespaced models for Role and Permission
* Updates User model to use HasRoles trait
* Adds email-preview for use while in local development
* Adds tenant seeder to seed an admin user.
* Updates the create command to accept a name and email

+++ app/Console/Commands/DeleteTenant.php
@@ -34,8 +34,15 @@ class DeleteTenant extends Command
         $force = $this->option('force');
         
         if ($tenant = $this->getTenantByFqdn($fqdn)) {
-            app(WebsiteRepository::class)->delete($tenant->website, $force);
-            $tenant->delete();
+            if ($tenant->website) {
+                app(WebsiteRepository::class)->delete($tenant->website, $force);
+            }
+            
+            app(HostnameRepository::class)->delete($tenant, $force);
+    
+            $this->info("Tenant {$fqdn} successfully deleted.");
+        } else {
+            $this->error("Could not find tenant {$fqdn}");
         }
     }
